+++ app/src/main/java/org/simple/clinic/home/overdue/search/OverdueSearchUi.kt
@@ -4,7 +4,6 @@ interface OverdueSearchUi {
   fun showSearchHistory(searchHistory: Set<String>)
   fun hideSearchResults()
   fun hideSearchHistory()
-  fun renderSearchQuery(searchQuery: String)
   fun hideProgress()
   fun showProgress()
   fun showSearchResults()