+++ app/src/main/java/org/simple/clinic/home/overdue/search/OverdueSearchUiRenderer.kt
@@ -10,7 +10,6 @@ class OverdueSearchUiRenderer(
 ) : ViewRenderer<OverdueSearchModel> {
 
   override fun render(model: OverdueSearchModel) {
-    ui.renderSearchQuery(model.searchQuery.orEmpty())
     ui.hideProgress()
 
     if (model.hasSearchQuery) {