+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - Wire Mobius loop with overdue search screen UI
 - Handle keyboard visibility in overdue search screen
 - Bump mixpanel to v6.3.0
+- Bump Google Services plugin to v4.3.12
 - [In Progress: 27 Apr 2022] Migrate `BloodPressureHistoryScreenEffectHandler` to use view effect
 
 ## 2022-06-20-8300