+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 
 ### Internal
 
+- [In Progress: 27 Apr 2022] Migrate `BloodPressureHistoryScreenEffectHandler` to use view effect
+
+## 2022-06-27-8307
+
+### Internal
+
 - Save search query to overdue search history when search query is valid
 - Add layout for overdue search screen
 - Bump Sentry to v6.1.2
@@ -28,7 +34,6 @@
 - Bump Google Services plugin to v4.3.12
 - Don't show no patients found error view when search query is empty
 - Don't show empty search history items in overdue search screen
-- [In Progress: 27 Apr 2022] Migrate `BloodPressureHistoryScreenEffectHandler` to use view effect
 
 ## 2022-06-20-8300