+++ CHANGELOG.md
@@ -6,10 +6,15 @@
 
 - Prefill search query when search history item is clicked
 - Highlight search results with the matched search query
-- Fix old synced appointments not getting deleted on purge
 - Bump datadog sdk to v1.13.0
 - Move existing search history item to the top when clicked
 - [In Progress: 27 Apr 2022] Migrate `BloodPressureHistoryScreenEffectHandler` to use view effect
+
+## 2022-06-27-8310
+
+### Internal
+
+- Fix old synced appointments not getting deleted on purge
 
 ## 2022-06-27-8307